+++ backend/src/main/java/com/minibeit/post/service/PostFileService.java
@@ -4,10 +4,15 @@ import com.minibeit.common.component.file.S3Uploader;
 import com.minibeit.common.dto.SavedFile;
 import com.minibeit.post.domain.Post;
 import com.minibeit.post.domain.PostFile;
+import com.minibeit.post.domain.repository.PostFileRepository;
+import com.minibeit.post.domain.repository.PostRepository;
+import com.minibeit.post.dto.PostRequest;
+import com.minibeit.post.dto.PostResponse;
+import com.minibeit.post.service.exception.PostNotFoundException;
+import com.minibeit.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,12 +23,21 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class PostFileService {
     private final S3Uploader s3Uploader;
+    private final PostFileRepository postFileRepository;
+    private final PostRepository postRepository;
+    private final PostPermissionCheck postPermissionCheck;
 
-    public List<PostFile> uploadFiles(Post post, List<MultipartFile> files) {
+    public PostResponse.OnlyId addFiles(Long postId, PostRequest.AddFile request, User user) {
+        Post post = postRepository.findByIdWithBusinessProfile(postId).orElseThrow(PostNotFoundException::new);
+        postPermissionCheck.userInBusinessProfileCheck(post.getBusinessProfile().getId(), user);
+
         List<SavedFile> savedFiles = new ArrayList<>();
-        if (files != null) {
-            savedFiles = s3Uploader.uploadFileList(files);
+        if (request.getFiles() != null) {
+            savedFiles = s3Uploader.uploadFileList(request.getFiles());
         }
-        return savedFiles.stream().map(postFile -> PostFile.create(post, postFile)).collect(Collectors.toList());
+        List<PostFile> postFiles = savedFiles.stream().map(postFile -> PostFile.create(post, postFile)).collect(Collectors.toList());
+        postFileRepository.saveAll(postFiles);
+
+        return PostResponse.OnlyId.build(post);
     }
 }
