+++ backend/src/rest/AuthController.http
@@ -2,7 +2,7 @@ POST {{apiUrl}}/user/login
 Content-Type: application/json;charset=UTF-8
 
 {
-  "id": "6"
+  "id": "1"
 }
 
 > {%
